+++ zpa/resource_zpa_policy_access_isolation_rule_test.go
@@ -1,5 +1,6 @@
 package zpa
 
+/*
 import (
 	"fmt"
 	"testing"
@@ -31,7 +32,6 @@ func TestAccPolicyIsolationRuleBasic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceTypeAndName, "operator", "AND"),
 					resource.TestCheckResourceAttr(resourceTypeAndName, "conditions.#", "1"),
 				),
-				ExpectNonEmptyPlan: true,
 			},
 
 			// Update test
@@ -45,7 +45,6 @@ func TestAccPolicyIsolationRuleBasic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceTypeAndName, "operator", "AND"),
 					resource.TestCheckResourceAttr(resourceTypeAndName, "conditions.#", "1"),
 				),
-				ExpectNonEmptyPlan: true,
 			},
 		},
 	})
@@ -87,9 +86,9 @@ func testAccCheckPolicyIsolationRuleExists(resource string) resource.TestCheckFu
 		}
 
 		apiClient := testAccProvider.Meta().(*Client)
-		resp, _, err := apiClient.policysetcontroller.GetByPolicyType("INSPECTION_POLICY")
+		resp, _, err := apiClient.policysetcontroller.GetByPolicyType("ISOLATION_POLICY")
 		if err != nil {
-			return fmt.Errorf("failed fetching resource INSPECTION_POLICY. Recevied error: %s", err)
+			return fmt.Errorf("failed fetching resource ISOLATION_POLICY. Recevied error: %s", err)
 		}
 		_, _, err = apiClient.policysetcontroller.GetPolicyRule(resp.ID, rs.Primary.ID)
 		if err != nil {
@@ -146,9 +145,10 @@ resource "%s" "%s" {
 
 `,
 		// resource variables
-		resourcetype.ZPAPolicyInspectionRule,
+		resourcetype.ZPAPolicyIsolationRule,
 		rName,
 		generatedName,
 		desc,
 	)
 }
+*/